+++ CHANGELOG.md
@@ -18,6 +18,8 @@
   - The resource no longer uses Win32_Product WMI class when evaluating if SQL Server Management Studio is installed. See article [kb974524](https://support.microsoft.com/en-us/kb/974524) for more information.
   - Now it uses CIM cmdlets to get information from WMI classes.
   - Resolved all of the PSScriptAnalyzer warnings that was triggered in the common tests.
+- Changes to xSQLServerScript
+  - All credential parameters now also has the type [System.Management.Automation.Credential()] to work with PowerShell 4.0.
 
 ## 4.0.0.0